refactor(ui actions): make ui.ts file entrypoint for imports and handle platform specific logic internally
