+++ app/actions/platformSpecific/ui.electron.ts
@@ -1,6 +1,5 @@
 import { ipcRenderer } from 'electron'
-import { UI_SIGNOUT } from '../reducers/ui.TARGET_PLATFORM'
-export * from './ui.base'
+import { UI_SIGNOUT } from '../../reducers/ui.TARGET_PLATFORM'
 
 export const signout = () => {
   ipcRenderer.send('block-menus', true)
