+++ app/actions/platformSpecific/ui.web.ts
@@ -1,0 +1,7 @@
+import { UI_SIGNOUT } from '../../reducers/ui.TARGET_PLATFORM'
+
+export const signout = () => {
+  return {
+    type: UI_SIGNOUT
+  }
+}
