--- app/actions/ui.web.ts
@@ -1,8 +1,0 @@
-import { UI_SIGNOUT } from '../reducers/ui.TARGET_PLATFORM'
-export * from './ui.base'
-
-export const signout = () => {
-  return {
-    type: UI_SIGNOUT
-  }
-}
