+++ app/components/modals/RemoveDataset.tsx
@@ -5,7 +5,7 @@ import { ApiAction } from '../../store/api'
 
 import { connectComponentToProps } from '../../utils/connectComponentToProps'
 
-import { dismissModal } from '../../actions/ui.TARGET_PLATFORM'
+import { dismissModal } from '../../actions/ui'
 import { removeDatasetAndFetch } from '../../actions/api'
 
 import { selectModal } from '../../selections'
